Fix rand_int type

+++ include/lattice.h
@@ -60,7 +60,7 @@ public:
     }
 
     double rand_int(int min, int max) {
-        std::uniform_int_distribution<double> dist(min, max);
+        std::uniform_int_distribution<int> dist(min, max);
         return dist(p.gen);
     }
 
@@ -133,7 +133,7 @@ public:
     }
 
     double rand_int(int min, int max) {
-        std::uniform_int_distribution<double> dist(min, max);
+        std::uniform_int_distribution<int> dist(min, max);
         return dist(p.gen);
     }
